Use symlinks over default-tiddler-location
I noticed that when using `tiddlywiki wiki --server` that instead of
tiddlers being modified when saving changes it started creating
sequentially numbered versions of the tiddlers.

For example after running the server and opening the page in the browser
I immediately had a new file added called
`wiki-dist/tiddlers/$__StoryList 1.tid`. Then if I stop and restarted
the server I'd end up with yet another additional file:
`wiki-dist/tiddlers/$__StoryList 2.tid`.

I guessed this was an artifact of using the
`config.default-tiddler-location` instead. It is too bad but it seems
the only way to have multiple additions share the same `tiddlers` folder
is to us symbolic links.